requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.8
django-bootstrap4==1.1.1
django-braces==1.14.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-registration==3.0.1
django-star-ratings==0.9.1
django-tinymce==2.8.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
swapper==1.1.2.post1
typed-ast==1.4.0
whitenoise==5.0.1
wrapt==1.11.2
